fix(setup): split locked AC expectations under 400-line domain test cap (#4054)
